decoy_service/requirements.txt
selenium>=4.15.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
faker>=20.0.0
